Initial Implementation of a portal system for Epochtalk
This system takes the latest 10 threads from a given board and display
the last post from each of the threads in order. This acts as a sort of
feed of activity for that board.

It also has a list of the 10 most recent posts at the bottom as a more
global feeds of activity.

Further updates could include an admin defined space for content and
maybe a news section.